rebase & cleanup
